Finished Implementing Evolution Mechanic & Created PokemonFactory
* Pokemon Class
Implemented justLeveledUp flag that keeps track of Pokemon that leveled up in the battle.
Created IV and EV array accessor
Accessors for individual skills
Nature accessor
Adjusted setHealthAndExp method to account for double exp value instead of int

* Created PokemonFactory
 Has a method to create evolutions.

* BattleState
Now checks at the end of the battle for Pokemon that just leveled up and evolves them if they are the correct level.